requirements.txt
# Audio Processing
pydub>=0.25.1
librosa>=0.10.1
ffmpeg-python>=0.2.0
soundfile>=0.12.1
scipy>=1.11.4
numpy>=1.26.0

# Deep Learning
torch>=2.1.0
torchaudio>=2.1.0
torchvision>=0.16.0
snorkel>=0.9.0

# Utilities & ML
scikit-learn>=1.3.0
pandas>=2.1.0
matplotlib>=3.8.0
tqdm>=4.66.0
seaborn>=0.13.0

# Optional but helpful
jupyterlab>=4.0.7
ipython>=8.18.0

# SWE
streamlit>=1.20.0
fastapi>=0.95.0
fastapi[standard]